pyproject.toml
[project]
name = "tt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "openai>=2.15.0",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
]
